+++ src/main/java/com/github/mikkiko/pockettrainer/entity/User.java
@@ -6,6 +6,11 @@ import lombok.experimental.Accessors;
 import javax.persistence.*;
 
 import java.util.List;
+import java.util.Set;
+
+/**
+ * User entity.
+ */
 
 @Accessors(chain = true)
 @Data
@@ -14,28 +19,34 @@ import java.util.List;
 public class User {
 
     @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Integer id;
+
+    @Column(unique = true)
     private String email;
 
     private String password;
 
-    private String name;
+    private String nickname;
 
     @OneToMany(cascade = CascadeType.ALL, fetch = FetchType.EAGER)
     @JoinColumn(name = "user_id")
     private List<Training> trainings;
 
+    @ElementCollection(targetClass = Role.class, fetch = FetchType.EAGER)
+    @CollectionTable(name = "user_roles", joinColumns = @JoinColumn(name = "user_id"))
+    @Enumerated(EnumType.STRING)
+    private Set<Role> roles;
+
     @Override
     public String toString() {
         return "User{" +
-                "email='" + email + '\'' +
+                "id=" + id +
+                ", email='" + email + '\'' +
                 ", password='" + password + '\'' +
-                ", trainings=" + (trainings.isEmpty() ? 0 : trainings) +
-                "}\n";
+                ", nickname='" + nickname + '\'' +
+                ", trainings=" + (trainings == null ? 0 : trainings) +
+                ", roles=" + (roles == null ? 0 : roles) +
+                '}';
     }
 }
-/*
-@GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
-
-    @Column(unique = true)
- */