+++ src/main/java/com/github/mikkiko/pockettrainer/exception/ExercisesException.java
@@ -1,0 +1,20 @@
+package com.github.mikkiko.pockettrainer.exception;
+
+/**
+ * Exercise exception extends {@link CustomException}.
+ */
+
+public class ExercisesException extends CustomException{
+
+    public ExercisesException(String message) {
+        super(message);
+    }
+
+    public ExercisesException(Cause cause) {
+        super(cause);
+    }
+
+    public ExercisesException(Cause cause, Integer id) {
+        super(cause, id);
+    }
+}