+++ src/main/java/com/github/mikkiko/pockettrainer/exception/NoSuchInfoException.java
@@ -1,4 +1,0 @@
-package com.github.mikkiko.pockettrainer.exception;
-
-public class NoSuchInfoException extends Exception{
-}