+++ src/main/java/com/github/mikkiko/pockettrainer/repository/ImageRepository.java
@@ -4,6 +4,10 @@ import com.github.mikkiko.pockettrainer.entity.Image;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+/**
+ * {@link Repository} for handling with {@link Image} entity.
+ */
+
 @Repository
 public interface ImageRepository extends JpaRepository<Image, Integer> {
 }