+++ src/main/java/com/github/mikkiko/pockettrainer/repository/UserRepository.java
@@ -7,10 +7,19 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
+import java.util.Optional;
+
+/**
+ * {@link Repository} for handling with {@link User} entity.
+ */
 
 @Repository
-public interface UserRepository extends JpaRepository<User, String> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
-    @Query("select u.trainings from User u where u.email like :email")
-    List<Training> getUserTrainings(String email);
+    @Query("select u.trainings from User u where u.id in :id")
+    List<Training> getUserTrainings(Integer id);
+
+    Optional<User> findUserByEmail(String email);
+
+    Boolean existsUserByEmail(String email);
 }