+++ src/main/java/com/github/mikkiko/pockettrainer/services/TrainingInfoServiceImpl.java
@@ -4,9 +4,9 @@ import com.github.mikkiko.pockettrainer.dto.TrainingInfoDTO;
 import com.github.mikkiko.pockettrainer.entity.Exercise;
 import com.github.mikkiko.pockettrainer.entity.Training;
 import com.github.mikkiko.pockettrainer.entity.TrainingInfo;
-import com.github.mikkiko.pockettrainer.exception.NoSuchExerciseException;
-import com.github.mikkiko.pockettrainer.exception.NoSuchInfoException;
-import com.github.mikkiko.pockettrainer.exception.NoSuchTrainingException;
+import com.github.mikkiko.pockettrainer.exception.Cause;
+import com.github.mikkiko.pockettrainer.exception.ExercisesException;
+import com.github.mikkiko.pockettrainer.exception.TrainingsException;
 import com.github.mikkiko.pockettrainer.repository.ExerciseRepository;
 import com.github.mikkiko.pockettrainer.repository.TrainingInfoRepository;
 import com.github.mikkiko.pockettrainer.repository.TrainingRepository;
@@ -14,8 +14,13 @@ import com.github.mikkiko.pockettrainer.util.EntityMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
+
+/**
+ * Implementation of {@link TrainingInfoService}.
+ */
 
 @Service
 @RequiredArgsConstructor
@@ -27,60 +32,44 @@ public class TrainingInfoServiceImpl implements TrainingInfoService {
     private final EntityMapper mapper;
 
     @Override
-    public boolean saveTrainingInfo(TrainingInfoDTO dto) {
-        try {
-            Exercise exercise = exerciseRepo.findById(dto.getExerciseId())
-                    .orElseThrow(NoSuchExerciseException::new);
-            Training training = trainingRepo.findById(dto.getTrainingId())
-                    .orElseThrow(NoSuchTrainingException::new);
+    public void saveTrainingInfo(TrainingInfoDTO dto) throws TrainingsException, ExercisesException {
+        Exercise exercise = exerciseRepo.findById(dto.getExerciseId()).orElseThrow(
+                () -> new ExercisesException(Cause.EXERCISE_NOT_FOUND, dto.getExerciseId()));
 
-            infoRepo.save(mapper.fromTrainingInfoDtoToEntity(dto, training, exercise));
-            return true;
-        } catch (NoSuchExerciseException | NoSuchTrainingException e) {
-            e.printStackTrace();
-        }
-        return false;
+        Training training = trainingRepo.findById(dto.getTrainingId()).orElseThrow(
+                () -> new TrainingsException(Cause.TRAINING_NOT_FOUND, dto.getTrainingId()));
+        infoRepo.save(mapper.fromTrainingInfoDtoToEntity(dto, training, exercise));
     }
 
     @Override
-    public TrainingInfoDTO getTrainingInfoById(Integer id) {
-        TrainingInfoDTO dto = null;
-        try {
-            TrainingInfo info = infoRepo.findById(id).orElseThrow(NoSuchInfoException::new);
-            dto = mapper.fromEntityToTrainingInfoDto(info);
-        } catch (NoSuchInfoException e) {
-            e.printStackTrace();
-        }
-        return dto;
+    public TrainingInfoDTO getTrainingInfoById(Integer id) throws TrainingsException {
+        TrainingInfo info = infoRepo.findById(id).orElseThrow(
+                () -> new TrainingsException(Cause.TRAINING_NOT_FOUND, id));
+        return mapper.fromEntityToTrainingInfoDto(info);
     }
 
     @Override
-    public List<TrainingInfoDTO> getTrainingsInfoList(List<Integer> ids) {
-        return ids
-                .stream()
-                .map(this::getTrainingInfoById)
-                .collect(Collectors.toList());
+    public List<TrainingInfoDTO> getTrainingsInfoList(List<Integer> infosID) throws TrainingsException {
+        List<TrainingInfoDTO> trainingInfoDTOList = new ArrayList<>();
+        for (Integer id : infosID)
+            trainingInfoDTOList.add(getTrainingInfoById(id));
+        return trainingInfoDTOList;
     }
 
     @Override
-    public boolean saveTrainingsInfoList(List<TrainingInfoDTO> dtoList) {
-        try {
-            if (!dtoList.isEmpty()) {
-                TrainingInfoDTO first = dtoList.get(0);
-                final Exercise exercise = exerciseRepo.findById(first.getExerciseId())
-                        .orElseThrow(NoSuchExerciseException::new);
-                final Training training = trainingRepo.findById(first.getTrainingId())
-                        .orElseThrow(NoSuchTrainingException::new);
-                List<TrainingInfo> entityList = dtoList
-                        .stream()
-                        .map(dto -> mapper.fromTrainingInfoDtoToEntity(dto, training, exercise))
-                        .collect(Collectors.toList());
-                entityList.forEach(infoRepo::save);
-                return true;
-            }
-        }catch (NoSuchExerciseException | NoSuchTrainingException e){
-            e.printStackTrace();
-        }
-        return false;
+    public void saveTrainingsInfoList(List<TrainingInfoDTO> trainingInfoDTOs)
+            throws TrainingsException, ExercisesException {
+        if (!trainingInfoDTOs.isEmpty()) {
+            TrainingInfoDTO first = trainingInfoDTOs.get(0);
+            final Exercise exercise = exerciseRepo.findById(first.getExerciseId()).orElseThrow(
+                    () -> new ExercisesException(Cause.EXERCISE_NOT_FOUND, first.getExerciseId()));
+            final Training training = trainingRepo.findById(first.getTrainingId()).orElseThrow(
+                    () -> new TrainingsException(Cause.TRAINING_NOT_FOUND, first.getTrainingId()));
+            List<TrainingInfo> entityList = trainingInfoDTOs
+                    .stream()
+                    .map(dto -> mapper.fromTrainingInfoDtoToEntity(dto, training, exercise))
+                    .collect(Collectors.toList());
+            entityList.forEach(infoRepo::save);
+        } else throw new TrainingsException(Cause.TRAINING_LIST_EMPTY);
     }
 }