+++ src/main/java/com/github/mikkiko/pockettrainer/services/UserService.java
@@ -2,28 +2,89 @@ package com.github.mikkiko.pockettrainer.services;
 
 import com.github.mikkiko.pockettrainer.dto.TrainingDTO;
 import com.github.mikkiko.pockettrainer.dto.UserDTO;
-import com.github.mikkiko.pockettrainer.entity.Training;
 import com.github.mikkiko.pockettrainer.entity.User;
-import com.github.mikkiko.pockettrainer.exception.NoSuchUserException;
+import com.github.mikkiko.pockettrainer.exception.UserException;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
 
+/**
+ * {@link Service} for handling {@link User} entity.
+ */
+
 public interface UserService {
 
-    Optional<User> getUser(String email);
+    /**
+     * Find {@link User} by ID.
+     *
+     * @param id provided User ID
+     * @return {@link User} with provided ID or null otherwise.
+     */
+    Optional<User> getUserById(Integer id);
 
+    /**
+     * Find {@link User} by email.
+     *
+     * @param email provided User email.
+     * @return {@link User} with provided email or null otherwise.
+     */
+    Optional<User> getUserByEmail(String email);
+
+    /**
+     * Save provided {@link User} entity.
+     *
+     * @param user provided user.
+     */
     boolean saveUser(User user);
 
-    boolean isUserExist(String email);
+    /**
+     * Check {@link User} entity existing.
+     *
+     * @param email provided User email.
+     * @return true if {@link User} with provided email exists or false in the other case.
+     */
+    boolean isUserExistByEmail(String email);
 
+    /**
+     * Retrieve all {@link User} entities.
+     *
+     * @return {@link List} of {@link User}.
+     */
     List<User> getAllUsers();
 
-    Optional<User> deleteUserByEmail(String email);
+    /**
+     * Delete {@link User} by ID.
+     *
+     * @param id provided User ID
+     * @return deleted {@link User} with provided ID or null otherwise.
+     */
+    Optional<User> deleteUserById(Integer id);
 
-    UserDTO getUserDTO(String email);
+    /**
+     * Get {@link UserDTO} by user ID.
+     *
+     * @param id provided User ID.
+     * @return {@link UserDTO} with provided ID or null otherwise.
+     * @throws UserException if {@link User} with provided id does not exist.
+     */
+    UserDTO getUserDTOById(Integer id) throws UserException;
 
-    boolean updateUserTrainings(String email, TrainingDTO training);
+    /**
+     * Update {@link User} trainings list.
+     *
+     * @param id provided User ID.
+     * @param training provided {@link TrainingDTO}.
+     * @throws UserException if {@link User} with provided id does not exist.
+     */
+    void updateUserTrainings(Integer id, TrainingDTO training) throws UserException;
 
-    boolean updatePassword(String email, String newPassword);
+    /**
+     * Update {@link User} password.
+     *
+     * @param id provided User ID.
+     * @param newPassword provided new password.
+     * @throws UserException if {@link User} with provided id does not exist.
+     */
+    void updatePassword(Integer id, String newPassword) throws UserException;
 }